+++ src/apps/loan-list/list/loan-list.entry.tsx
@@ -7,31 +7,26 @@ import withIsPatronBlockedHoc from "../../../core/utils/withIsPatronBlockedHoc";
 import { BlockedPatronEntryTextProps } from "../../../core/storybook/blockedArgs";
 import { withConfig } from "../../../core/utils/config";
 import GlobalUrlEntryPropsInterface from "../../../core/utils/types/global-url-props";
+import { GroupModalProps } from "../../../core/storybook/groupModalArgs";
+import { GroupModalLoansProps } from "../../../core/storybook/loanGroupModalArgs";
+import { MaterialDetailsModalProps } from "../../../core/storybook/materialDetailsModalArgs";
 
 export interface LoanListEntryConfigProps {
   thresholdConfig: string;
 }
 export interface LoanListEntryUrlProps {
-  loanListEreolenUrl: string;
   materialOverdueUrl: string;
   thresholdConfig: string;
 }
 
 interface LoanListEntryTextProps {
-  groupModalDueDateLinkToPageWithFeesText: string;
-  groupModalDueDateHeaderText: string;
-  groupModalDueDateWarningLoanOverdueText: string;
   loanListAriaLabelListButtonText: string;
   loanListAriaLabelStackButtonText: string;
-  groupModalRenewLoanDeniedInterLibraryLoanText: string;
-  groupModalRenewLoanDeniedMaxRenewalsReachedText: string;
-  groupModalRenewLoanDeniedReservedText: string;
   loanListDigitalLoansEmptyListText: string;
   loanListDigitalLoansTitleText: string;
   loanListDigitalPhysicalLoansEmptyListText: string;
   loanListDueDateModalAriaLabelText: string;
   loanListDueDateModalAriaDescribeMobileText: string;
-  groupModalHiddenLabelCheckboxOnMaterialText: string;
   loanListLateFeeDesktopText: string;
   loanListLateFeeMobileText: string;
   loanListMaterialDaysText: string;
@@ -44,48 +39,27 @@ interface LoanListEntryTextProps {
   loanListStatusBadgeWarningText: string;
   loanListTitleText: string;
   loanListToBeDeliveredDigitalMaterialText: string;
-  groupModalDueDateMaterialText: string;
   loanListToBeDeliveredText: string;
   materialAndAuthorText: string;
   materialByAuthorText: string;
-  materialDetailsCloseModalAriaLabelText: string;
-  materialDetailsPhysicalDueDateLabelText: string;
-  materialDetailsLinkToPageWithFeesText: string;
-  materialDetailsLoanDateLabelText: string;
-  materialDetailsMaterialNumberLabelText: string;
-  materialDetailsModalAriaDescriptionText: string;
-  materialDetailsOverdueText: string;
-  materialDetailsRenewLoanButtonText: string;
-  materialDetailsWarningLoanOverdueText: string;
   publizonAudioBookText: string;
   publizonEbookText: string;
   publizonPodcastText: string;
-  groupModalAriaDescriptionText: string;
-  groupModalButtonText: string;
-  groupModalCheckboxText: string;
-  groupModalCloseModalAriaLabelText: string;
-  groupModalHeaderText: string;
   resultPagerStatusText: string;
-  showMoreText: string;
-  groupModalReturnLibraryText: string;
-  materialDetailsGoToEreolenText: string;
   loanListMaterialLateFeeText: string;
   loanListMaterialDayText: string;
   loanListStatusCircleAriaLabelText: string;
-  materialDetailsDigitalDueDateLabelText: string;
-  groupModalGoToMaterialText: string;
-  groupModalGoToMaterialAriaLabelText: string;
 }
 
 export interface LoanListEntryWithPageSizeProps
   extends BlockedPatronEntryTextProps,
     LoanListEntryTextProps,
     LoanListEntryConfigProps,
+    GroupModalProps,
+    GroupModalLoansProps,
     LoanListEntryUrlProps,
-    GlobalUrlEntryPropsInterface {
-  pageSizeDesktop?: number;
-  pageSizeMobile?: number;
-}
+    MaterialDetailsModalProps,
+    GlobalUrlEntryPropsInterface {}
 
 const LoanListEntry: FC<LoanListEntryWithPageSizeProps> = ({
   pageSizeDesktop,
