+++ src/apps/menu/menu.entry.tsx
@@ -4,6 +4,13 @@ import { withText } from "../../core/utils/text";
 import { withConfig } from "../../core/utils/config";
 import { withUrls } from "../../core/utils/url";
 import GlobalUrlEntryPropsInterface from "../../core/utils/types/global-url-props";
+import { GroupModalProps } from "../../core/storybook/groupModalArgs";
+import { GroupModalLoansProps } from "../../core/storybook/loanGroupModalArgs";
+import { GroupModalReservationsProps } from "../../core/storybook/reservationGroupModalArgs";
+import { pageSizeGlobal } from "../../core/utils/helpers/general";
+import { MaterialDetailsModalProps } from "../../core/storybook/materialDetailsModalArgs";
+import { ReservationMaterialDetailsProps } from "../../core/storybook/reservationMaterialDetailsArgs";
+import { DeleteReservationModalArgs } from "../../core/storybook/deleteReservationModalArgs";
 
 export interface MenuProps {
   menuViewYourProfileText: string;
@@ -22,8 +29,8 @@ export interface MenuProps {
   menuLoginText: string;
   menuLoginUrl: string;
   menuSignUpText: string;
+  reservationsReadyText: string;
   menuSignUpUrl: string;
-  menuNotificationsMenuAriaLabelText: string;
   menuProfileLinksAriaLabelText: string;
   menuUserIconAriaLabelText: string;
   menuNotAuthenticatedCloseButtonText: string;
@@ -32,6 +39,25 @@ export interface MenuProps {
   menuNotAuthenticatedModalDescriptionText: string;
 }
 
-const MenuEntry: FC<MenuProps & GlobalUrlEntryPropsInterface> = () => <Menu />;
+const MenuEntry: FC<
+  MenuProps &
+    GlobalUrlEntryPropsInterface &
+    GroupModalProps &
+    GroupModalLoansProps &
+    DeleteReservationModalArgs &
+    GroupModalReservationsProps &
+    MaterialDetailsModalProps &
+    ReservationMaterialDetailsProps
+> = ({ pageSizeDesktop, pageSizeMobile }) => {
+  const pageSize = pageSizeGlobal(
+    {
+      desktop: pageSizeDesktop,
+      mobile: pageSizeMobile
+    },
+    "pageSizeLoanList"
+  );
+
+  return <Menu pageSize={pageSize} />;
+};
 
 export default withUrls(withConfig(withText(MenuEntry)));
