+++ src/apps/reservation-list/list/reservation-list.dev.tsx
@@ -5,6 +5,8 @@ import ReservationList from "./reservation-list.entry";
 import { getModalIds } from "../../../core/utils/helpers/general";
 import serviceUrlArgs from "../../../core/storybook/serviceUrlArgs";
 import blockedArgs from "../../../core/storybook/blockedArgs";
+import reservationMaterialDetailsProps from "../../../core/storybook/reservationMaterialDetailsArgs";
+import deleteReservationModalArgs from "../../../core/storybook/deleteReservationModalArgs";
 
 export default {
   title: "Apps / Reservation list",
@@ -12,6 +14,8 @@ export default {
   argTypes: {
     // Config
     ...serviceUrlArgs,
+    ...deleteReservationModalArgs,
+    ...reservationMaterialDetailsProps,
     ...blockedArgs,
     blacklistedSearchBranchesConfig: {
       name: "Blacklisted branches",
@@ -36,11 +40,6 @@ export default {
       name: "Branches",
       defaultValue:
         '[\n   {\n      "branchId":"DK-775120",\n      "title":"Højbjerg"\n   },\n   {\n      "branchId":"DK-775122",\n      "title":"Beder-Malling"\n   },\n   {\n      "branchId":"DK-775144",\n      "title":"Gellerup"\n   },\n   {\n      "branchId":"DK-775167",\n      "title":"Lystrup"\n   },\n   {\n      "branchId":"DK-775146",\n      "title":"Harlev"\n   },\n   {\n      "branchId":"DK-775168",\n      "title":"Skødstrup"\n   },\n   {\n      "branchId":"FBS-751010",\n      "title":"Arresten"\n   },\n   {\n      "branchId":"DK-775147",\n      "title":"Hasle"\n   },\n   {\n      "branchId":"FBS-751032",\n      "title":"Må ikke benyttes"\n   },\n   {\n      "branchId":"FBS-751031",\n      "title":"Fjernlager 1"\n   },\n   {\n      "branchId":"DK-775126",\n      "title":"Solbjerg"\n   },\n   {\n      "branchId":"FBS-751030",\n      "title":"ITK"\n   },\n   {\n      "branchId":"DK-775149",\n      "title":"Sabro"\n   },\n   {\n      "branchId":"DK-775127",\n      "title":"Tranbjerg"\n   },\n   {\n      "branchId":"DK-775160",\n      "title":"Risskov"\n   },\n   {\n      "branchId":"DK-775162",\n      "title":"Hjortshøj"\n   },\n   {\n      "branchId":"DK-775140",\n      "title":"Åby"\n   },\n   {\n      "branchId":"FBS-751009",\n      "title":"Fjernlager 2"\n   },\n   {\n      "branchId":"FBS-751029",\n      "title":"Stadsarkivet"\n   },\n   {\n      "branchId":"FBS-751027",\n      "title":"Intern"\n   },\n   {\n      "branchId":"FBS-751026",\n      "title":"Fælles undervejs"\n   },\n   {\n      "branchId":"FBS-751025",\n      "title":"Fællessekretariatet"\n   },\n   {\n      "branchId":"DK-775133",\n      "title":"Bavnehøj"\n   },\n   {\n      "branchId":"FBS-751024",\n      "title":"Fjernlånte materialer"\n   },\n   {\n      "branchId":"DK-775100",\n      "title":"Hovedbiblioteket"\n   },\n   {\n      "branchId":"DK-775170",\n      "title":"Trige"\n   },\n   {\n      "branchId":"DK-775150",\n      "title":"Tilst"\n   },\n   {\n      "branchId":"DK-775130",\n      "title":"Viby"\n   },\n   {\n      "branchId":"DK-775164",\n      "title":"Egå"\n   }\n]',
-      control: { type: "text" }
-    },
-    interestPeriodsConfig: {
-      defaultValue:
-        '[\n   {\n      "value":"30",\n      "label":"1 month"\n   },\n   {\n      "value":"60",\n      "label":"2 months"\n   },\n   {\n      "value":"90",\n      "label":"3 months"\n   },\n   {\n      "value":"180",\n      "label":"6 months"\n   },\n   {\n      "value":"360",\n      "label":"1 year"\n   }\n]',
       control: { type: "text" }
     },
     // Page size
@@ -151,10 +150,6 @@ export default {
       defaultValue: "day",
       control: { type: "text" }
     },
-    reservationDetailsButtonRemoveText: {
-      defaultValue: "Remove your reservation",
-      control: { type: "text" }
-    },
     reservationDetailsOthersInQueueText: {
       defaultValue: "Others are queueing for this material",
       control: { type: "text" }
@@ -163,56 +158,12 @@ export default {
       defaultValue: "@count queued",
       control: { type: "text" }
     },
-    reservationDetailsStatusTitleText: {
-      defaultValue: "Status",
-      control: { type: "text" }
-    },
     reservationDetailsExpiresTitleText: {
       defaultValue: "Pickup deadline",
       control: { type: "text" }
     },
     reservationDetailsDigitalMaterialExpiresTitleText: {
       defaultValue: "Borrow before",
-      control: { type: "text" }
-    },
-    reservationDetailsPickUpAtTitleText: {
-      defaultValue: "Pickup branch",
-      control: { type: "text" }
-    },
-    reservationDetailsChangeText: {
-      defaultValue: "Apply changes",
-      control: { type: "text" }
-    },
-    reservationDetailsPickupDeadlineTitleText: {
-      defaultValue: "Pickup deadline",
-      control: { type: "text" }
-    },
-    reservationDetailsDateOfReservationTitleText: {
-      defaultValue: "Date of reservation",
-      control: { type: "text" }
-    },
-    reservationDetailsNoInterestAfterTitleText: {
-      defaultValue: "Not interested after",
-      control: { type: "text" }
-    },
-    reservationDetailsReadyForLoanText: {
-      defaultValue: "Ready for pickup",
-      control: { type: "text" }
-    },
-    reservationDetailsRemoveDigitalReservationText: {
-      defaultValue: "Remove your reservation",
-      control: { type: "text" }
-    },
-    reservationDetailAllowRemoveReadyReservationsConfig: {
-      defaultValue: "1",
-      control: { type: "text" }
-    },
-    reservationDetailsDigitalReservationGoToEreolenText: {
-      defaultValue: "Go to eReolen",
-      control: { type: "text" }
-    },
-    reservationDetailsBorrowBeforeText: {
-      defaultValue: "Borrow before @date",
       control: { type: "text" }
     },
     reservationDetailsExpiresText: {
@@ -225,34 +176,6 @@ export default {
     },
     reservationDetailsCancelText: {
       defaultValue: "Cancel",
-      control: { type: "text" }
-    },
-    deleteReservationModalHeaderText: {
-      defaultValue:
-        '{"type":"plural","text":["Cancel reservation","Cancel reservations"]}',
-      control: { type: "text" }
-    },
-    deleteReservationModalDeleteQuestionText: {
-      defaultValue:
-        '{"type":"plural","text":["Do you want to cancel your reservation?","Do you want to cancel your reservations?"]}',
-      control: { type: "text" }
-    },
-    deleteReservationModalNotRegrettableText: {
-      defaultValue: "You cannot regret this action",
-      control: { type: "text" }
-    },
-    deleteReservationModalDeleteButtonText: {
-      defaultValue:
-        '{"type":"plural","text":["Cancel reservation","Cancel reservations"]}',
-      control: { type: "text" }
-    },
-    deleteReservationModalCloseModalText: {
-      defaultValue: "Close delete reservation modal",
-      control: { type: "text" }
-    },
-    deleteReservationModalAriaDescriptionText: {
-      defaultValue:
-        "This button opens a modal that covers the entire page and contains the possibility to delete a selected reservation, or multiple selected reservations",
       control: { type: "text" }
     },
     reservationListPauseReservationText: {
